backend/requirements.txt
reportlab
pandas
openpyxl
flask
flask-cors
pillow
reportlab
